SEQNG-724: Moved all common commands from GPIClient to GiapiClient.
