+++ modules/giapi/src/main/scala/giapi/client/GiapiClient.scala
@@ -9,16 +9,56 @@ import giapi.client.commands.{Command, CommandResult, Configuration}
 
 import scala.concurrent.duration._
 
+/////////////////////////////////////////////////////////////////
+// The GiapiClient comprises the common commands for such clients
+/////////////////////////////////////////////////////////////////
 trait GiapiClient[F[_]] {
+  import GiapiClient.DefaultCommandTimeout
+
   def giapi: Giapi[F]
 
-  def genericApply(configuration: Configuration): F[CommandResult] =
+  def test: F[CommandResult] =
     giapi.command(
-      Command(
-        SequenceCommand.APPLY,
+      Command(SequenceCommand.TEST, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def init: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.INIT, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def datum: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.DATUM, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def park: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.PARK, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def verify: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.VERIFY,
         Activity.PRESET_START,
-        configuration
-      ), GiapiClient.DefaultCommandTimeout)
+        Configuration.Zero), DefaultCommandTimeout)
+
+  def endVerify: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.END_VERIFY,
+        Activity.PRESET_START,
+        Configuration.Zero), DefaultCommandTimeout)
+
+  def guide: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.GUIDE, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def endGuide: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.END_GUIDE,
+        Activity.PRESET_START,
+        Configuration.Zero), DefaultCommandTimeout)
 
   def observe[A: Show](dataLabel: A, expTime: FiniteDuration): F[CommandResult] =
     giapi.command(
@@ -27,6 +67,41 @@ trait GiapiClient[F[_]] {
         Activity.PRESET_START,
         Configuration.single(commands.DataLabelCfg, dataLabel)
       ), expTime)
+
+  def endObserve: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.END_OBSERVE,
+        Activity.PRESET_START,
+        Configuration.Zero), DefaultCommandTimeout)
+
+  def pause: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.PAUSE, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def continue: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.CONTINUE,
+        Activity.PRESET_START,
+        Configuration.Zero), DefaultCommandTimeout)
+
+  def stop: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.STOP, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def abort: F[CommandResult] =
+    giapi.command(
+      Command(SequenceCommand.ABORT, Activity.PRESET_START, Configuration.Zero),
+      DefaultCommandTimeout)
+
+  def genericApply(configuration: Configuration): F[CommandResult] =
+    giapi.command(
+      Command(
+        SequenceCommand.APPLY,
+        Activity.PRESET_START,
+        configuration
+      ), DefaultCommandTimeout)
 }
 
 object GiapiClient {
