+++ modules/giapi/src/main/scala/giapi/client/gpi/GPIClient.scala
@@ -37,79 +37,6 @@ class GPIClient[F[_]](override val giapi: Giapi[F]) extends GiapiClient[F] {
   def heartbeatS: F[Stream[F, Int]] =
     giapi.stream[Int]("gpi:heartbeat")
 
-  ///////////////////
-  // General commands
-  ///////////////////
-  def test: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.TEST, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def init: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.INIT, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def datum: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.DATUM, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def park: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.PARK, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def verify: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.VERIFY,
-              Activity.PRESET_START,
-              Configuration.Zero), DefaultCommandTimeout)
-
-  def endVerify: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.END_VERIFY,
-              Activity.PRESET_START,
-              Configuration.Zero), DefaultCommandTimeout)
-
-  def guide: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.GUIDE, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def endGuide: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.END_GUIDE,
-              Activity.PRESET_START,
-              Configuration.Zero), DefaultCommandTimeout)
-
-  def endObserve: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.END_OBSERVE,
-              Activity.PRESET_START,
-              Configuration.Zero), DefaultCommandTimeout)
-
-  def pause: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.PAUSE, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def continue: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.CONTINUE,
-              Activity.PRESET_START,
-              Configuration.Zero), DefaultCommandTimeout)
-
-  def stop: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.STOP, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
-  def abort: F[CommandResult] =
-    giapi.command(
-      Command(SequenceCommand.ABORT, Activity.PRESET_START, Configuration.Zero),
-      DefaultCommandTimeout)
-
   ////////////////////////
   // GPI Specific commands
   ////////////////////////
